+++ capsule/src/main/java/ca/allanwang/capsule/library/interfaces/CCollapseListener.java
@@ -16,19 +16,16 @@ public abstract class CCollapseListener implements AppBarLayout.OnOffsetChangedL
     public final void onOffsetChanged(AppBarLayout appBarLayout, int verticalOffset) {
         onVerticalOffsetChanged(appBarLayout, verticalOffset);
         if (verticalOffset == 0) {
-            if (mCurrentState != State.EXPANDED) {
+            if (mCurrentState != State.EXPANDED)
                 onStateChanged(appBarLayout, State.EXPANDED);
-            }
             mCurrentState = State.EXPANDED;
         } else if (Math.abs(verticalOffset) >= appBarLayout.getTotalScrollRange()) {
-            if (mCurrentState != State.COLLAPSED) {
+            if (mCurrentState != State.COLLAPSED)
                 onStateChanged(appBarLayout, State.COLLAPSED);
-            }
             mCurrentState = State.COLLAPSED;
         } else {
-            if (mCurrentState != State.IDLE) {
+            if (mCurrentState != State.IDLE)
                 onStateChanged(appBarLayout, State.IDLE);
-            }
             mCurrentState = State.IDLE;
         }
     }