+++ capsule/src/main/java/ca/allanwang/capsule/library/swiperecyclerview/animators/FadeInAnimator.java
@@ -1,0 +1,49 @@
+package ca.allanwang.capsule.library.swiperecyclerview.animators;
+
+import android.support.v4.view.ViewCompat;
+import android.support.v7.widget.RecyclerView;
+import android.view.animation.Interpolator;
+
+import ca.allanwang.capsule.library.swiperecyclerview.wasabeef.animators.BaseItemAnimator;
+
+
+/**
+ * Created by Allan Wang on 2017-03-18.
+ */
+
+public class FadeInAnimator extends BaseItemAnimator {
+
+    public FadeInAnimator() {
+    }
+
+    public FadeInAnimator(Interpolator interpolator) {
+        mInterpolator = interpolator;
+    }
+
+    @Override
+    protected void animateRemoveImpl(final RecyclerView.ViewHolder holder) {
+        ViewCompat.animate(holder.itemView)
+                .alpha(0)
+                .setDuration(getRemoveDuration())
+                .setInterpolator(mInterpolator)
+                .setListener(new DefaultRemoveVpaListener(holder))
+                .setStartDelay(getRemoveDelay(holder))
+                .start();
+    }
+
+    @Override
+    protected void preAnimateAddImpl(RecyclerView.ViewHolder holder) {
+        ViewCompat.setAlpha(holder.itemView, 0);
+    }
+
+    @Override
+    protected void animateAddImpl(final RecyclerView.ViewHolder holder) {
+        ViewCompat.animate(holder.itemView)
+                .alpha(1)
+                .setDuration(getAddDuration())
+                .setInterpolator(mInterpolator)
+                .setListener(new DefaultAddVpaListener(holder))
+                .setStartDelay(getAddDelay(holder))
+                .start();
+    }
+}