+++ capsule/src/main/java/ca/allanwang/capsule/library/swiperecyclerview/interfaces/FinishCallback.java
@@ -1,0 +1,8 @@
+package ca.allanwang.capsule.library.swiperecyclerview.interfaces;
+
+/**
+ * Created by Allan Wang on 2017-03-09.
+ */
+
+public interface FinishCallback {
+}